requirements.txt
boto3==1.12.39
botocore==1.15.49
docutils==0.15.2
jmespath==0.9.5
python-dateutil==2.8.1
pyyaml==5.3.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9 ; python_version != '3.4'
